pyproject.toml
[project]
name = "prisoners-dilemma-simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.1",
    "networkx>=3.4.2",
    "omegaconf>=2.3.0",
    "openai>=1.66.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "vllm>=0.7.3",
]
